Increase hash rounds
Provide some very basic protection for dictionary attacks. A better solution is to create ephemeral keys that do not depend on the password to prevent passive offline attacks on the handshake. Active MITM attack resistance will not be improved except by stronger passwords, more hash rounds, or the use of a certificate to protect the connection based on the ephemeral keys.